pyproject.toml
[tool.poetry]
name = "boaviztapi"
version = "1.3.12"
description = "An API to access Boavizta's methodologies and footprint reference data"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
uvicorn = "^0.22"
pandas = "^2.2.3"
aiofile = "^3.8"
numpy = "^2.2.0"
scipy = "^1.14.1"
rapidfuzz = "^3.0"
markdown = "^3.4"
mangum = "^0.17"
importlib-metadata = "^6.6.0"
pyyaml = "^6.0"
toml = "^0.10.2"
fastapi-cache2 = "^0.2.2"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
xmltodict = "^1.0.2"
google-auth = "^2.41.1"
python-multipart = "^0.0.20"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
atomicwrites = "^1.4.1"
httpx = "^0.28.1"
requests = "^2.32.3"
jsonschema = "^4.25.1"
pycountry = "^24.6.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.4"
mkdocs-render-swagger-plugin = "^0.1.0"
mkdocs-macros-plugin = "^1.0.4"
linkcheckmd = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
